RR-116 - Refactor function that returns latest functional skills
