+++ server/routes/overview/educationAndTrainingView.ts
@@ -1,5 +1,4 @@
 import type { FunctionalSkills, PrisonerSummary } from 'viewModels'
-import { mostRecentFunctionalSkills } from './mappers/functionalSkillsMapper'
 
 export default class EducationAndTrainingView {
   constructor(private readonly prisonerSummary: PrisonerSummary, private readonly functionalSkills: FunctionalSkills) {}
@@ -12,7 +11,7 @@ export default class EducationAndTrainingView {
     return {
       tab: 'education-and-training',
       prisonerSummary: this.prisonerSummary,
-      functionalSkills: mostRecentFunctionalSkills(this.functionalSkills),
+      functionalSkills: this.functionalSkills,
     }
   }
 }
